Remomved a warning in pio_conf_pin_to_peripheral().

+++ src/sam3x8e/pio.h
@@ -309,6 +309,6 @@ uint32_t pio_read_port(pio_reg_t *port);
  * @bug {Not yet tested. All ports will be included after initial testing.}
  */
 uint8_t pio_conf_pin_to_peripheral(pio_reg_t *port,
-		uint32_t periph, uint32_t pin_number);
+		uint32_t periph, uint8_t pin_number);
 
 #endif
